+++ utils/rudderstack.ts
@@ -2,24 +2,24 @@ import { getWindow } from 'utils/ssr';
 
 declare global {
   interface Window {
-    analytics: {
+    rudderanalytics: {
       track: (event: string, properties: { [key: string]: string | boolean }) => void;
       page: (category?: string, name?: string) => void;
     };
   }
 }
 
-// Segment track
+// RudderStack track
 export const track = (event: string, properties: { [key: string]: string | boolean }) => {
   const environment = getWindow();
-  if (environment && environment.analytics) {
-    environment.analytics.track(event, properties);
+  if (environment && environment.rudderanalytics) {
+    environment.rudderanalytics.track(event, properties);
   }
 };
 
 export const page = (category?: string, name?: string) => {
   const environment = getWindow();
-  if (environment && environment.analytics) {
-    environment.analytics.page(category, name);
+  if (environment && environment.rudderanalytics) {
+    environment.rudderanalytics.page(category, name);
   }
 };
